refactor(story-engine-repository): split RepositoryV2 into domain repos and keep RepositoryOld as commit wrapper
- Add domain-specific repositories: `ProjectRepository`, `StoryRepository`, `CharacterRepository`, and `EditEventRepository`.
- Convert `RepositoryV2` into a small composition root that exposes domain repositories.
- Update `RepositoryOld` to delegate to domain repos while preserving legacy `commit/refresh` behavior.
- Export `RepositoryV2` from `repository/__init__.py` for direct transaction-oriented usage.
- Wire `Service` and `ImageUploadService` to use `RepositoryV2` for transaction-scoped edit-event operations without breaking existing external `Repository` consumers.
