+++ backend/core/story_engine/repository/__init__.py
@@ -1,7 +1,9 @@
 from .exception import NotFoundError
 from .repository_old import RepositoryOld as Repository
+from .repository_v2 import RepositoryV2
 
 __all__ = [
     "Repository",
     "NotFoundError",
+    "RepositoryV2",
 ]
