+++ backend/core/story_engine/repository/repository_v2.py
@@ -1,242 +1,15 @@
-import uuid
-from typing import Any
-
-from slugify import slugify
-from sqlalchemy import desc, func, select
 from sqlalchemy.ext.asyncio import AsyncSession
-from sqlalchemy.orm import joinedload, selectinload
 
-from ..models import Character, EditEvent, Project, Story
-from ..models.edit_event import EditEventStatus, OperationType
-from .exception import NotFoundError
+from .character_repository import CharacterRepository
+from .edit_event_repository import EditEventRepository
+from .project_repository import ProjectRepository
+from .story_repository import StoryRepository
 
 
 class RepositoryV2:
     def __init__(self, db: AsyncSession):
         self.db = db
-
-    async def get_all_projects_of_user_with_story_count(
-        self, user_id: str
-    ) -> list[tuple[Project, int]]:
-        query = (
-            select(Project, func.count(Story.id).label("story_count"))
-            .outerjoin(Story, Story.project_id == Project.id)
-            .where(Project.user_id == uuid.UUID(user_id))
-            .group_by(Project.id)
-        )
-        result = await self.db.execute(query)
-        projects = result.tuples().all()
-        return list(projects)
-
-    async def create_project(self, user_id: str, name: str | None = None) -> Project:
-        project = Project(user_id=uuid.UUID(user_id), name=name)
-        self.db.add(project)
-        return project
-
-    async def get_project_details(
-        self, user_id: str, project_id: uuid.UUID
-    ) -> Project | None:
-        query = (
-            select(Project)
-            .where(Project.id == project_id, Project.user_id == uuid.UUID(user_id))
-            .options(
-                selectinload(Project.stories),
-            )
-        )
-        result = await self.db.execute(query)
-        return result.scalar_one_or_none()
-
-    async def create_new_story(self, project_id: uuid.UUID) -> Story:
-        story = Story(project_id=project_id)
-        self.db.add(story)
-        return story
-
-    async def get_story(
-        self, project_id: uuid.UUID, story_id: uuid.UUID
-    ) -> Story | None:
-        stmt = select(Story).where(Story.id == story_id, Story.project_id == project_id)
-        result = await self.db.execute(stmt)
-        return result.scalar_one_or_none()
-
-    async def delete_story(self, project_id: uuid.UUID, story_id: uuid.UUID) -> None:
-        story = await self.get_story(project_id, story_id)
-        if story is None:
-            raise NotFoundError(f"Story {story_id} not found in project {project_id}")
-        await self.db.delete(story)
-
-    async def get_story_with_project(self, story_id: uuid.UUID) -> Story | None:
-        stmt = (
-            select(Story).where(Story.id == story_id).options(joinedload(Story.project))
-        )
-        result = await self.db.execute(stmt)
-        return result.scalar_one_or_none()
-
-    async def update_story_with_story_text_and_user_input_text(
-        self,
-        story_id: uuid.UUID,
-        story_text: str,
-        user_input_text: str,
-        source_event_id: uuid.UUID | None = None,
-    ) -> Story:
-        story = await self.db.get(Story, story_id)
-        if story is None:
-            raise NotFoundError(f"Story {story_id} not found")
-
-        story.story_text = story_text
-        story.user_input_text = user_input_text
-        if source_event_id is not None:
-            story.source_event_id = source_event_id
-        return story
-
-    async def create_edit_event(
-        self,
-        project_id: uuid.UUID,
-        target_type: str,
-        target_id: uuid.UUID,
-        operation_type: OperationType,
-        user_instruction: str,
-        input_snapshot: dict[str, Any] | None = None,
-    ) -> EditEvent:
-        event = EditEvent(
-            project_id=project_id,
-            target_type=target_type,
-            target_id=target_id,
-            operation_type=operation_type.value,
-            user_instruction=user_instruction,
-            input_snapshot=input_snapshot,
-        )
-        self.db.add(event)
-        return event
-
-    async def update_edit_event(
-        self,
-        edit_event_id: uuid.UUID,
-        status: EditEventStatus,
-        output_snapshot: dict[str, Any] | None = None,
-    ) -> EditEvent:
-        event = await self.db.get(EditEvent, edit_event_id)
-        if event is None:
-            raise NotFoundError(f"EditEvent {edit_event_id} not found")
-
-        event.status = status.value
-        if output_snapshot is not None:
-            event.output_snapshot = output_snapshot
-        return event
-
-    async def get_edit_events_for_target(
-        self, target_type: str, target_id: uuid.UUID, limit: int = 20
-    ) -> list[EditEvent]:
-        stmt = (
-            select(EditEvent)
-            .where(
-                EditEvent.target_type == target_type,
-                EditEvent.target_id == target_id,
-            )
-            .order_by(desc(EditEvent.created_at))
-            .limit(limit)
-        )
-        result = await self.db.execute(stmt)
-        events = result.scalars().all()
-        return list(events)
-
-    async def bulk_create_characters(self, characters: list[Character]) -> None:
-        self.db.add_all(characters)
-
-    async def get_all_characters_for_a_story(
-        self, story_id: uuid.UUID
-    ) -> list[Character]:
-        stmt = select(Character).where(Character.story_id == story_id)
-        result = await self.db.execute(stmt)
-        characters = result.scalars().all()
-        return list(characters)
-
-    async def get_character(
-        self, character_id: uuid.UUID, story_id: uuid.UUID
-    ) -> Character | None:
-        stmt = select(Character).where(
-            Character.id == character_id, Character.story_id == story_id
-        )
-        result = await self.db.execute(stmt)
-        return result.scalar_one_or_none()
-
-    async def get_character_for_user_in_project_and_story(
-        self,
-        user_id: str,
-        project_id: uuid.UUID,
-        story_id: uuid.UUID,
-        character_id: uuid.UUID,
-    ) -> Character | None:
-        stmt = (
-            select(Character)
-            .join(Story)
-            .join(Project)
-            .where(
-                Character.id == character_id,
-                Story.id == story_id,
-                Project.id == project_id,
-                Project.user_id == uuid.UUID(user_id),
-            )
-        )
-        result = await self.db.execute(stmt)
-        return result.scalar_one_or_none()
-
-    async def update_character(
-        self, character_id: uuid.UUID, story_id: uuid.UUID, updates: dict[str, Any]
-    ) -> Character:
-        character = await self.get_character(character_id, story_id)
-        if character is None:
-            raise NotFoundError(
-                f"Character {character_id} not found in story {story_id}"
-            )
-
-        meta = updates.pop("meta", None)
-        attribute_updates = {k: v for k, v in updates.items() if v is not None}
-        if attribute_updates:
-            character.attributes = {**character.attributes, **attribute_updates}
-            if "name" in attribute_updates:
-                character.name = attribute_updates["name"]
-
-        if meta is not None:
-            character.meta = {**character.meta, **meta}
-        return character
-
-    async def replace_character_attributes(
-        self,
-        character_id: uuid.UUID,
-        story_id: uuid.UUID,
-        attributes: dict[str, Any],
-        source_event_id: uuid.UUID | None = None,
-    ) -> Character:
-        character = await self.get_character(character_id, story_id)
-        if character is None:
-            raise NotFoundError(
-                f"Character {character_id} not found in story {story_id}"
-            )
-
-        character.attributes = attributes
-        character.name = attributes["name"]
-        character.slug = slugify(attributes["name"])
-        if source_event_id is not None:
-            character.source_event_id = source_event_id
-        return character
-
-    async def delete_character(
-        self, character_id: uuid.UUID, story_id: uuid.UUID
-    ) -> None:
-        character = await self.get_character(character_id, story_id)
-        if character is None:
-            raise NotFoundError(
-                f"Character {character_id} not found in story {story_id}"
-            )
-        await self.db.delete(character)
-
-    async def update_character_render_url(
-        self, character_id: uuid.UUID, story_id: uuid.UUID, render_url: str
-    ) -> Character:
-        character = await self.get_character(character_id, story_id)
-        if character is None:
-            raise NotFoundError(
-                f"Character {character_id} not found in story {story_id}"
-            )
-        character.render_url = render_url
-        return character
+        self.project = ProjectRepository(db)
+        self.story = StoryRepository(db)
+        self.character = CharacterRepository(db)
+        self.edit_event = EditEventRepository(db)
